requirements.txt
matplotlib
progressbar2
flask
waitress
gunicorn==20.0.4